lambda/requirements-rest.txt
requests==2.31.0
google-auth==2.23.4